Ajaxの学習。 いいねボタンのJavaScript化が完成した。 次に移る
