+++ app/javascript/packs/application.js
@@ -4,7 +4,7 @@
 // that code so it'll be compiled.
 
 require("@rails/ujs").start()
-require("turbolinks").start()
+// require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
 
@@ -18,68 +18,3 @@ require("channels")
 
 require("trix")
 require("@rails/actiontext")
-
-import $ from 'jquery'
-import axios from 'axios'
-import 'trix/dist/trix.css'
-import { csrfToken } from 'rails-ujs'
-
-axios.defaults.headers.common['X-CSRF-Token'] = csrfToken()
-
-
-document.addEventListener('turbolinks:load', () => {
-    const dataset = document.getElementById('article-show').dataset;
-    const articleId = dataset.articleId;
-    axios.get(`/articles/${articleId}/like`)
-      .then((response) => {
-         const hasLiked = response.data.hasLiked
-         if (hasLiked) {
-           $('.active-heart').removeClass('hidden')
-         } else {
-           $('.inactive-heart').removeClass('hidden')
-         }
-      })
- })
-
-const handleHeartDisplay = (hasLiked) => {
-    if (hasLiked) {
-      $('.active-heart').removeClass('hidden')
-    } else {
-      $('.inactive-heart').removeClass('hidden')
-    }
-}
-
-document.addEventListener('turbolinks:load', () => {
-    const dataset = document.getElementById('article-show').dataset;
-    const articleId = dataset.articleId
-
-    axios.get(`/articles/${articleId}/like`)
-      .then((response) => {
-         const hasLiked = response.data.hasLiked
-         handleHeartDisplay(hasLiked)
-    })
-
-    $('.inactive-heart').on('click', () => {
-      axios.post(`/articles/${articleId}/like`)
-        .then((response) => {
-          console.log(response)
-        })
-        .catch((e) => {
-          window.alert('Error')
-          console.log(e)
-        })
-    })   
-
-    $('.active-heart').on('click', () => {
-        axios.delete(`/articles/${articleId}/like`)
-          .then((response) => {
-            console.log(response)
-          })
-          .catch((e) => {
-            window.alert('Error')
-            console.log(e)
-          })
-      })   
-
- })
-
