+++ app/views/articles/show.html.haml
@@ -33,3 +33,5 @@
   = link_to new_article_comment_path(@article) do
     .btn-secondary
       コメントを追加
+
+= javascript_pack_tag 'article'
